Redefine OMX_U32 on 64-bit architecture
-Redefine OMX_U32 to unsigned int on 64-bit architecture.

+++ mm-core/inc/OMX_Types.h
@@ -142,10 +142,10 @@ typedef unsigned short OMX_U16;
 typedef signed short OMX_S16;
 
 /** OMX_U32 is a 32 bit unsigned quantity that is 32 bit word aligned */
-typedef unsigned long OMX_U32;
+typedef unsigned int OMX_U32;
 
 /** OMX_S32 is a 32 bit signed quantity that is 32 bit word aligned */
-typedef signed long OMX_S32;
+typedef signed int OMX_S32;
 
 
 /* Users with compilers that cannot accept the "long long" designation should
